almost finished with recursive insert
Chances are that the OS crashes anyway, in case a malloc or calloc
issue is encountered, but this is quite the exercise. Just have to add
a recoverymode from the memmove and memcpy, that happen if a node is
added. After this, the library should fail gracefully if it can't find
the necessary memory.