I got to a point where debugging is possible
Fixed pointers and the tests
Currently test 2 fails because the insert operation is doing something
wrong. Duplicate inserts are not detected, which is comical, because
that was the point of this whole exercise. Detecting that while still
beeing able to detect the failure of the insert function, when e.g
encountering memory allocation errors.